add multiLayer architure
